+++ src/theme/colors.ts
@@ -1,39 +1,3 @@
-import { extendTheme } from '@chakra-ui/react';
-import { Trirong, Snippet, Work_Sans } from 'next/font/google'
-
-
-/** font families */
-
-const work_sans = Work_Sans({
-  weight: ['100', '200', '300', '400', '500', '600', '700', '800', '900'],
-  style: ['normal', 'italic'],
-  preload: true,
-  fallback: ['sans-serif'],
-  subsets: ['latin', 'latin-ext']
-});
-const snippet = Snippet({
-  weight: '400',
-  style: 'normal',
-  preload: true,
-  fallback: ['monospace'],
-  subsets: ['latin']
-});
-const trigong = Trirong({
-  weight: ['100', '200', '300', '400', '500', '600', '700', '800', '900'],
-  style: ['normal', 'italic'],
-  preload: true,
-  fallback: ['serif'],
-  subsets: ['latin', 'latin-ext']
-});
-
-const fonts = {
-  text: work_sans.style.fontFamily,
-  heading: snippet.style.fontFamily,
-  display: trigong.style.fontFamily,
-}
-
-/** application colors */
-
 const colors = {
 
   /** base palette */
@@ -115,32 +79,4 @@ const colors = {
 
 };
 
-
-/** application styles */
-
-const styles = {
-  global: {
-    body: {
-      fontFamily: 'text',
-      color: 'light',
-      bg: 'dark',
-    },
-    a: {
-      textDecoration: 'none',
-      fontFamily: 'heading',
-    },
-    button: {
-      fontFamily: 'heading',
-    }
-  },
-}
-
-
-/** component variants */
-
-const components = {}
-
-
-/** export theme */
-
-export const theme = extendTheme({ fonts, colors, styles, components });
+export default colors;
