+++ src/theme/components.ts
@@ -1,0 +1,5 @@
+
+
+const components = {}
+
+export default components;
