+++ src/theme/styles.ts
@@ -1,0 +1,20 @@
+
+
+const styles = {
+  global: {
+    body: {
+      fontFamily: 'text',
+      color: 'light',
+      bg: 'dark',
+    },
+    a: {
+      textDecoration: 'none',
+      fontFamily: 'heading',
+    },
+    button: {
+      fontFamily: 'heading',
+    }
+  },
+}
+
+export default styles;
